+++ src/main/java/com/github/simiacryptus/aicoder/Icons.java
@@ -5,7 +5,6 @@ import com.intellij.openapi.util.IconLoader;
 import javax.swing.*;
 
 public class Icons {
-
   public static final Icon icon1 = IconLoader.getIcon("/icons/sdk_16.svg", Icons.class);
   public static final Icon icon2 = IconLoader.getIcon("/icons/edit.svg", Icons.class);